+++ GoalsAndHabits/app/src/androidTest/java/cmput301f17t09/goalsandhabits/NewProfileActivityTest.java
@@ -6,6 +6,8 @@ import android.view.View;
 
 import com.robotium.solo.Solo;
 
+import cmput301f17t09.goalsandhabits.Profiles.NewProfileActivity;
+
 /**
  * Created by Andrew on 10/22/2017.
  */
@@ -13,7 +15,7 @@ import com.robotium.solo.Solo;
 public class NewProfileActivityTest extends ActivityInstrumentationTestCase2<NewProfileActivity>{
     private Solo solo;
 
-    public NewProfileActivityTest() {super(cmput301f17t09.goalsandhabits.NewProfileActivity.class);}
+    public NewProfileActivityTest() {super(NewProfileActivity.class);}
 
     public void setUp() throws Exception{
         solo = new Solo(getInstrumentation(), getActivity());
@@ -28,6 +30,8 @@ public class NewProfileActivityTest extends ActivityInstrumentationTestCase2<New
         View submitButton = getActivity().findViewById(R.id.submitButton);
         solo.clickOnView(submitButton);
         //TODO: Enter a username and save the profile to the database
+        //assert empty submission
+        solo.assertCurrentActivity("Wrong Activity", NewProfileActivity.class);
     }
 
     public void tearDown() throws Exception{
