+++ GoalsAndHabits/app/src/main/AndroidManifest.xml
@@ -9,28 +9,27 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".Main_Habits.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".FollowActivity" />
-        <activity android:name=".MapFiltersActivity" />
+        <activity android:name=".Follow_Activity.FollowActivity" />
+        <activity android:name=".Maps.MapFiltersActivity" />
         <activity
-            android:name=".HabitHistoryActivity"/>
-        <activity android:name=".NewHabit" />
-        <activity android:name=".ViewHabitActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".SignUpActivity" />
-        <activity android:name=".ProfileActivity" />
-        <activity android:name=".NewProfileActivity" />
+            android:name=".Main_Habits.HabitHistoryActivity"/>
+        <activity android:name=".Main_Habits.NewHabit" />
+        <activity android:name=".Main_Habits.ViewHabitActivity" />
+        <activity android:name=".Profiles.LoginActivity" />
+        <activity android:name=".Profiles.ProfileActivity" />
+        <activity android:name=".Profiles.NewProfileActivity" />
         <activity
-            android:name=".followerRequestsActivity"
-            android:parentActivityName=".ProfileActivity" />
-        <activity android:name=".MyHabitHistory"
-            android:parentActivityName=".ProfileActivity"></activity>
+            android:name=".Profiles.followerRequestsActivity"
+            android:parentActivityName=".Profiles.ProfileActivity" />
+        <activity android:name=".Profiles.MyHabitHistory"
+            android:parentActivityName=".Profiles.ProfileActivity"></activity>
     </application>
 
 </manifest>
