+++ GoalsAndHabits/app/src/main/java/cmput301f17t09/goalsandhabits/Main_Habits/HabitEvent.java
@@ -1,4 +1,4 @@
-package cmput301f17t09.goalsandhabits;
+package cmput301f17t09.goalsandhabits.Main_Habits;
 
 import java.io.Serializable;
 import java.util.Date;
