--- GoalsAndHabits/app/src/main/java/cmput301f17t09/goalsandhabits/MyHabitHistory.java
@@ -1,13 +1,0 @@
-package cmput301f17t09.goalsandhabits;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class MyHabitHistory extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_my_habit_history);
-    }
-}
