+++ GoalsAndHabits/app/src/main/java/cmput301f17t09/goalsandhabits/Profiles/MyHabitHistory.java
@@ -1,18 +1,19 @@
-package cmput301f17t09.goalsandhabits;
+package cmput301f17t09.goalsandhabits.Profiles;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 
-public class followerRequestsActivity extends AppCompatActivity {
+import cmput301f17t09.goalsandhabits.R;
+
+public class MyHabitHistory extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_follower_requests);
+        setContentView(R.layout.activity_my_habit_history);
 
         Toolbar toolbar = (Toolbar) findViewById(R.id.actionbar);
-        toolbar.setTitle("My Follower Requests");
-        setSupportActionBar(toolbar);
+        toolbar.setTitle("My Habit History");
     }
 }
