+++ GoalsAndHabits/app/src/main/java/cmput301f17t09/goalsandhabits/Profiles/followerRequestsActivity.java
@@ -1,0 +1,20 @@
+package cmput301f17t09.goalsandhabits.Profiles;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.support.v7.widget.Toolbar;
+
+import cmput301f17t09.goalsandhabits.R;
+
+public class followerRequestsActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_follower_requests);
+
+        Toolbar toolbar = (Toolbar) findViewById(R.id.actionbar);
+        toolbar.setTitle("My Follower Requests");
+        setSupportActionBar(toolbar);
+    }
+}
