--- GoalsAndHabits/app/src/main/java/cmput301f17t09/goalsandhabits/SignUpActivity.java
@@ -1,13 +1,0 @@
-package cmput301f17t09.goalsandhabits;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class SignUpActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sign_up);
-    }
-}
