+++ GoalsAndHabits/app/src/main/res/layout/activity_habit_history.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cmput301f17t09.goalsandhabits.HabitHistoryActivity">
+    tools:context="cmput301f17t09.goalsandhabits.Main_Habits.HabitHistoryActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/actionbar"
