+++ GoalsAndHabits/app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="8dp"
-    tools:context="cmput301f17t09.goalsandhabits.LoginActivity">
+    tools:context="cmput301f17t09.goalsandhabits.Profiles.LoginActivity">
