+++ GoalsAndHabits/app/src/main/res/layout/activity_map_filters.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cmput301f17t09.goalsandhabits.MapFiltersActivity"
+    tools:context="cmput301f17t09.goalsandhabits.Maps.MapFiltersActivity"
     android:label="This is the MapFiltersActivity">
 
     <android.support.v7.widget.Toolbar
